requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
